feat: Block remote input until connection established
This has some hacky signal-ish objects that represent the connection, whatever
its state.  This is handy because we can subscribe to it before it's connected
and react as its state changes.

The connection state does not degrade even if it gets disconnected.  Seems the
event listeners for the underlying connections aren't there yet.